+++ src/presenter/trip.js
@@ -1,0 +1,8 @@
+
+
+
+export default class Trip {
+  constructor(tripContainer) {
+    this._tripContainer = tripContainer;
+  }
+}
