+++ src/utils/render.js
@@ -1,0 +1,47 @@
+import Abstract from "../view/abstract";
+import {POSITION} from "../constans";
+
+export const renderElement = (container, element, place) => {
+  if (container instanceof Abstract) {
+    container = container.getElement();
+  }
+
+  if (element instanceof Abstract) {
+    element = element.getElement();
+  }
+
+  switch (place) {
+    case POSITION.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case POSITION.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
+
+export const replace = (newChild, oldChild) => {
+  if (newChild instanceof Abstract) {
+    newChild = newChild.getElement();
+  }
+
+  if (oldChild instanceof Abstract) {
+    oldChild = oldChild.getElement();
+  }
+
+  const parent = oldChild.parentElement;
+
+  if (parent === null || newChild === null || oldChild === null) {
+    throw new Error(`Can't replace unexisting elements`);
+  }
+
+  parent.replaceChild(newChild, oldChild);
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
